增加配置文件生成操作

+++ jeecg-boot/jeecg-boot-module-system/src/main/java/org/jeecg/modules/game/entity/GameChannel.java
@@ -56,7 +56,7 @@ public class GameChannel extends BaseEntity {
      */
     @Excel(name = "公告id", width = 15)
     @ApiModelProperty(value = "公告id")
-    private java.lang.Integer noticeId;
+    private java.lang.Long noticeId;
 
     /**
      * 版本号
